backend/requirements.txt
django>=2.1.6
djangorestframework>=3.9.1
django-rest-auth==0.9.3
django-allauth==0.37.1
django-phonenumber-field
phonenumbers
django-rest-swagger
django-cors-headers